Handle invalid (legacy) cache data
